have to fix some bud
