--- modeladmin/feedler/entry/index.html
@@ -1,11 +1,0 @@
-{% extends "modeladmin/index.html %}
-
-{% block header_extra %}
-  {% if user_can_create %}
-      <div class="right">
-          <div class="addbutton">
-            {% include 'modeladmin/includes/button.html' with button=view.button_helper.export_button %}
-           </div>
-       </div>
-   {% endif %}
-{% endblock %}
