+++ publichealth/home/templates/home/photo_gallery.html
@@ -1,0 +1,25 @@
+{% load compress static wagtailcore_tags wagtailimages_tags %}
+
+<!-- Gallery Template -->
+
+<div id="responsive" class="carousel-inner carousel-gallery slick" role="listbox">
+{% for block in page.gallery %}
+  {% for val in block.value %}
+		<div class="item">
+			{% image val.image fill-700x400 as entry_thumb %}
+			{% image val.image original as entry_photo %}
+			<image data-src="{{ entry_photo.url }}" data-caption="{{ val.caption }}" style="background-image:url({{ entry_thumb.url }})">
+		</div>
+  {% endfor %}
+{% endfor %}
+</div>
+<div class="carousel-inner slider-nav">
+{% for block in page.gallery %}
+  {% for val in block.value %}
+		<div class="item">
+			{% image val.image fill-700x400 as entry_thumb %}
+			<image data-src="{{ entry_thumb.url }}" data-caption="{{ val.caption }}" style="background-image:url({{ entry_thumb.url }})">
+		</div>
+  {% endfor %}
+{% endfor %}
+</div>
